+++ src/test/java/kitchenpos/table/application/OrderTableServiceTest.java
@@ -11,12 +11,12 @@ import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.order.request.OrderTableCreateRequest;
 import kitchenpos.support.ServiceTest;
 import kitchenpos.table.OrderTable;
 import kitchenpos.table.TableGroup;
 import kitchenpos.table.persistence.OrderTableRepository;
 import kitchenpos.table.persistence.TableGroupRepository;
+import kitchenpos.table.request.OrderTableCreateRequest;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Nested;
